requirements.txt
httpx>=0.24.0
python-telegram-bot[job-queue]>=21.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
psycopg2-binary>=2.9.0
